src/agentscope/workstation/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==3.2.0
billiard==4.2.1
celery==5.5.2
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
fastapi-limiter==0.1.6
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
kombu==5.5.3
loguru==0.6.0
passlib==1.7.4
pip==25.1
prompt-toolkit==3.0.51
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.9.1
pyjwt==2.8.0
pymysql==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
redis==6.0.0
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
pycryptodome==3.22.0
jieba==0.42.1

yaml
networkx

pymupdf==1.25.5
docx2txt==0.9
sandboxai-client==0.0.2
# pythonmonkey==1.1.0
agentdev==0.5.0
agentscope==0.1.5
weaviate-client==4.14.3
llama-index==0.12.34
llama-index-vector-stores-weaviate==1.3.1
llama-index-embeddings-dashscope==0.3.0
llama-index-postprocessor-dashscope-rerank==0.3.0
llama-index-readers-dashscope
llama-index-indices-managed-dashscope
# Markdown==3.8
# pypdfium2==4.30.1
llama-index-readers-file==0.4.7

